% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-cat.r
\name{geom_categorical}
\alias{geom_categorical}
\title{Sparkline layer for each map point}
\usage{
geom_categorical(mapping = aes(), data = NULL, stat = "sf",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{}

\item{data}{}

\item{stat}{}

\item{position}{}

\item{na.rm}{}

\item{show.legend}{}

\item{inherit.aes}{}

\item{...}{}
}
\description{
Plots a minimalist time series line chart at each coordinate
provided
}
\details{
.. content for \details{} ..
}
