% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-sparkline.R
\name{geom_sparkline}
\alias{geom_sparkline}
\title{Sparkline layer for each map point}
\usage{
geom_sparkline(mapping = aes(), data = NULL, stat = "sf",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\description{
Plots a minimalist time series line chart at each coordinate
provided
}
\details{
.. content for \details{} ..
}
